pyproject.toml
[tool.poetry]
name = "sf-course-proj"
version = "0.1.0"
description = ""
authors = ["Mark Chalenko <markus.chalenko@gmail.con>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["uvicorn"], version = "^0.111.0"}
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
greenlet = "^3.0.3"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
python-jose = {extras = ["cryptogrphy"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.1.3"
oauth = "^1.0.1"
setuptools = "^70.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
